Refactor : 장소 업로드 - 장소 선택 부분 refactoring
장소 선택 부분을 두 부분으로 나누어서 코드의 수를 줄이고 가독성을 높임
